requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
twilio==8.7.0
urllib3==2.0.4
yarl==1.9.2
